# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text or GIFs and
    easily stay in touch through your private chats or group discussion.
  version: "0.1"
servers:
  - url: http://localhost:3000
...
paths:
  /session:
    post:
      tags:
        - Login
      summary: Logs in or creates the user account
      security: []
      description: |
        Logs the user in by username. If the user does not exist, the account will be created, 
        and a unique identifier will be returned. If the user already exists, only the identifier 
        is returned, without creating a new account.
      operationId: doLogin
      requestBody:
        description: User's name is required for login or account creation.
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "The username used for login or registration."
              properties:
                username:
                  $ref: "#/components/schemas/Name"
              required:
                - username

      responses:
        "201":
          description: "User logged in successfully."
          content:
            application/json:
              schema:
                description: "Returns the user's unique identifier."
                type: object
                properties:
                  identifier:
                    $ref: "#/components/schemas/Identifier"
                required:
                  - identifier
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/conversationId"
    get:
      parameters:
        - name: limit
          in: query
          required: true
          description: Maximum number of a messages to return
          schema:
            type: integer
            example: 20
        - name: cursor
          in: query
          required: false
          description: |
            Token to retrieve the next set of messages starting after this cursor.
            not required because you can get your first set of data (messages).
          schema:
            $ref: "#/components/schemas/Identifier"
      tags:
        - Messages
      summary: list messages of a conversation
      description: |
        open a conversation to view all exchanged messages, displayed in reverse chronologial order.
        Each message include the timestamp, the content and the sender's username. and one/two checkmarks to indicate the status of sent messages.
      operationId: getConversation
      responses:
        "200":
          description: List of messages returned successfully
          content:
            application/json:
              schema:
                description: object with message details
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageDetail"
                    minItems: 0
                    maxItems: 250
                  totalCount:
                    type: integer
                    description: Total number of messages
                  nextCursor:
                    $ref: "#/components/schemas/Identifier"
                required:
                  - messages
                  - totalCount
                  - nextCursor
        "400":
          $ref: '#/components/responses/BadRequest'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/text:
    post:
      tags:
        - Messages
      summary: send a text message
      description: send a text message as a string to a spcific conversation using conversationId
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        description: needed a text to be sent
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTextMessage"
      responses:
        "202":
          description: message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetail"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/image:
    post:
      tags:
        - Messages
      summary: send an image
      description: send an image message, png or jpeg
      operationId: sendImage
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        description: image file needed
        required: true
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/Image"
          image/jpeg:
            schema:
              $ref: "#/components/schemas/Image"
      responses:
        "202":
          description: image sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetail"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/member:
    put:
      tags:
        - Groups
      summary: add an user in a group
      description: add a user in a group, you must be in the group already
      operationId: addToGroup
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: needed a userId to add a person in a group
              type: object
              properties:
                userId:
                  $ref: "#/components/schemas/Identifier"
              required:
                - userId
      responses:
        "201":
          description: users added
          content:
            application/json:
              schema:
                properties:
                  userId:
                    $ref: "#/components/schemas/Identifier"
                required:
                  - userId
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}:
    delete:
      tags:
        - Groups
      summary: leave a group
      description: leave a group using conversationId
      operationId: leaveGroup
      parameters:
        - $ref: '#/components/parameters/conversationId'
      responses:
        "204":
          description: group successfully abandoned
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/read:
    put:
      tags:
        - Messages
      summary: notify that the message has been read
      description: send a put request to notify the message with messageId has been read by user
      operationId: readMessage
      parameters:
        - $ref: "#/components/parameters/messageId"
      responses:
        "200":
          description: message status has been stored, return the message identifier
          content:
            application/json:
              schema:
                properties:
                  messageId:
                    $ref: "#/components/schemas/Identifier"
                required:
                  - messageId
        "400":
          $ref: '#/components/responses/BadRequest'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
...
    conversationId:
      description: identifier for a single conversation
      name: conversationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    messageId:
      description: identifier for a single message
      name: messageId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
  schemas:
    Name:
      title: UserName/GroupName
      description: schema used to rappresent username or group name
      type: string
      pattern: "^[a-zA-Z0-9]*$"
      minLength: 3
      maxLength: 25
      example: "Deadpool69"
    Identifier:
      title: ID
      description: "A unique identifier for the user, in the specified format."
      type: string
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      minLength: 15
      maxLength: 15
      example: "123e4567-e89b-12d3-a456-426614174000"
    Image:
      title: Image
      description: Image binary format as a string
      type: string
      format: binary
      minLength: 1
      maxLength: 20971520 # 20MB
...
    Text:
      description: body of the message
      type: string
      minLength: 1
      maxLength: 5000
      pattern: "^.*?$"
      example: "heyb buddy!"
    MessageDetail:
      description: message information
      type: object
      properties:
        messageId:
          $ref: "#/components/schemas/Identifier"
        senderId:
          $ref: "#/components/schemas/Identifier"
        username:
          description: sender of the message
          type: string
          example: "Deadpool"
          pattern: "^[a-zA-Z0-9]*$"
        text:
          $ref: "#/components/schemas/Text"
        image:
          description: "photo code (unique)"
          type: string
          minLength: 27
          maxLength: 27
          pattern: "^[0-9a-zA-Z]*?"
          example: "1ThDPX91uDwB1yxw5jYRxUIVs1W"
        replyTo:
          $ref: "#/components/schemas/Identifier"
        timestamp:
          description: data and time of message
          type: string
          format: date-time
          minLength: 20
          maxLength: 20
        status:
          description: status od the message
          type: string
          enum: [read,received]
        typeContent:
          description: content type, it can be a phot or a text
          type: string
          enum: [text,photo]
        reactions:
          description: array of reactions (emoji)
          type: array
          items:
            type: object
            description: object with the emoji idetifier and the sender id
            properties:
              emoji:
                description: user can react with one of this emoji
                type: string
                example: "😂"
                enum:
                  - "😊"
                  - "😂"
                  - "😍"
                  - "😎"
                  - "🥺"
              userId:
                $ref: "#/components/schemas/Identifier"
          minItems: 0
          maxItems: 250
      required:
        - messageId
        - senderId
        - username
        - text
        - date
        - status
        - typeContent
        - reactions
    NewTextMessage:
      description: field of a text message
      type: object
      properties:
        senderId:
          $ref: "#/components/schemas/Identifier"
        bodyMessage:
          $ref: "#/components/schemas/Text"
        replyTo:
          $ref: "#/components/schemas/Identifier"
      required:
        - senderId
        - bodyMessage
        - replyTo
  responses:
...
    BadRequest:
      description: syntax of the request was not compliant with docs, check parameters
      content:
        text/plain:
          schema:
            description: bad request string message
            type: string
            example: Bad Request
    #401
    Unauthorized:
      description: authorization missing or not valid
      content:
        text/plain:
          schema:
            description: Unauthorized string message
            type: string
            example: Unauthorized
    #403
    Forbidden:
      description: server will not accept request due to lacking of client permissions
      content:
        text/plain:
          schema:
            description: Forbidden string message
            type: string
            example: Forbidden
    #404
    NotFound:
      description: resource not found
      content:
        text/plain:
          schema:
            description: Not Found string message
            type: string
            example: Not found
    #409
    Conflict:
      description: resource cause a conflict
      content:
        text/plain:
          schema:
            description: conflict string message
            type: string
            example: Conflict
    #500 class responses
    #500
    InternalServerError:
      description: the server encountered an error while processing the request
      content:
        text/plain:
          schema:
            description: Internal Server Error string message
            type: string
            example: Internal Server Error
  securitySchemes:
    TokenAuth:
      description: |
        User authentication with the token generated by backend application
      type: http
      scheme: bearer

# Apply security scheme globally, disabling it explicitly when unnecessary with security: [].
security:
  - TokenAuth: []